LF-15813 usb: xhci-imx8m: Apply TXFIFO workaround only for iMX95 A0/A1
The workaround applied by commit d154606fc1af
(LF-11882-2 usb: xhci-imx8m: add workaround for USB3 host controller TXFIFO IC issue)
actually changes the TXFIFO memory address, it is only valid for iMX95 A1/A0
with the IC issue. On iMX95 B0, the issue is fixed. The changed memory space
is not valid, so will cause error when usb write.

To compatible with both iXM95 A1 and B0, update the driver to check
the SOC revision for applying the workaround.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Xu Yang <xu.yang_2@nxp.com>